CODE-RUB: Moving exceptions into correct folder
