+++ ISL.Providers.Captcha.GoogleReCaptcha/Models/Services/Foundations/Captcha/Exceptions/CaptchaDependencyValidationException.cs
@@ -4,11 +4,11 @@
 
 using Xeptions;
 
-namespace ISL.Providers.Captcha.GoogleReCaptcha.Models.Services.Foundations.Captcha
+namespace ISL.Providers.Captcha.GoogleReCaptcha.Models.Services.Foundations.Captcha.Exceptions
 {
-    public class CaptchaValidationException : Xeption
+    public class CaptchaDependencyValidationException : Xeption
     {
-        public CaptchaValidationException(string message, Xeption innerException)
+        public CaptchaDependencyValidationException(string message, Xeption innerException)
             : base(message, innerException)
         { }
     }
