+++ ISL.Providers.Captcha.GoogleReCaptcha/Services/Foundations/Captcha/CaptchaService.Exceptions.cs
@@ -2,7 +2,6 @@
 // Copyright (c) North East London ICB. All rights reserved.
 // ---------------------------------------------------------
 
-using ISL.Providers.Captcha.GoogleReCaptcha.Models.Services.Foundations.Captcha;
 using ISL.Providers.Captcha.GoogleReCaptcha.Models.Services.Foundations.Captcha.Exceptions;
 using System;
 using System.Threading.Tasks;
